requirements.txt
requests
python-dotenv
openai==0.28
beautifulsoup4


